+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **Added:** Install pre-commit. Run black, ruff, flake8 on pre-commit.
 - **Added:** Apply pre-commit badge to README.
 - **Added:** text.multireplace
+- **Improved:** iterator: splitter, uniques, uniquer, consume, all_equals, all_identical, all_equal, has_only_one
 
 ## v2.0.0 (2022-12-22) ##